+++ src/js/event-bus/index.js
@@ -1,0 +1,20 @@
+/*
+ * This file is part of the Front Foundation package.
+ *
+ * Copyright (c) 2017-present LIN3S <info@lin3s.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author Mikel Tuesta <mikeltuesta@gmail.com>
+ */
+
+import GMapInitializedEventSubscriber from './GMap/GMapInitializedEventSubscriber';
+import GMapMarkerSelectedEventSubscriber from './GMap/GMapMarkerSelectedEventSubscriber';
+import GMapGeocodeNoResultsEventSubscriber from './GMap/GMapGeocodeNoResultsEventSubscriber';
+
+export {
+  GMapInitializedEventSubscriber,
+  GMapMarkerSelectedEventSubscriber,
+  GMapGeocodeNoResultsEventSubscriber
+}