+++ tests/app/index.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * This file is part of the Front Foundation package.
+ *
+ * Copyright (c) 2017-present LIN3S <info@lin3s.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+use Symfony\Component\HttpFoundation\Request;
+
+require_once __DIR__ . '/vendor/autoload.php';
+$kernel = new Symfony('dev', true);
+$request = Request::createFromGlobals();
+$response = $kernel->handle($request);
+$response->send();
+$kernel->terminate($request, $response);